A better version on the "aternative method" of preemtive context swithing. This one actually works (if you skip the last part where R7 is popped back). Once again, checked in to seve as a reminder if the new method I have in mind turns hout to have drawbacks.
This method has one major drawback (even if I get it to work), and that is that
validation of CPU context after switching back is very hars and close to
impossible. The principle of the code checked in now should be clear, but
at least one register is NOT restored properlly - and that is PSW. That could
prolly be handled the same way as R7, byt the more complicated this mechanism
gets, the harder it is to validate - it takes hours just to cover one iteration
(and you need "a lot" of iterations to be even fairlly confident).