Introduced a new component for better and safer useage of the heap. Package is called \red KMEM and the files are tk_mem.c and tk_mem.h (so far).
Started to take care of the long needed issue with error codes and
better error handling. Introduced errno.h to begin with, whitch is part
of the package \ref kernel_reimpl_ansi. Its not a good solution yet,
since both kernel and ANSI codes are in the same file we have to invent
a way to omit the ANSI defines when a tool-chain that has errno.h is
used.