Found the nasty \ref BUG_000_001. Solution is robust but potentially degrades tinkers timing presition.
Found another bug caused by wraparound, that occures once every 71.6 minuts
but selcom shows itself.

@note  that systimer (sys_mickey) wraps once every 49.7 days but
kernel native time-keeping wraps 1000 times more often (71.6
minutes). This is due to that current precision on sys_time is in
mS, but kernel precision is in uS as a preparation to that the
\ref clock will be replaced by a higher precision time function
(\uptime or something similar).

prepared for better presision clock (true uS presision).