<b>Please read the in depth comments</b> about this check-in at \ref Blog051125
The test program (test.c) in this check-in is also particularly nasty
since it gives really long latencies on each task switch (up to and
above 500mS!). Test against this if you make any changes in either
timing or dispatching. Even considering this hard case, the drift was
very minor. The timing constants need re-trimming though.