- Replaced printf with printk in in-depth parts of the kernel. This is to make porting easier since printk can then be mapped to whatever counsole output ability there is (including none if there isn't any).
- Conditionals for: 1) time ISR latency and 2) clock systimer faliure
introduced. This is because debugging involves stopping the program but
not the clock HW, which will trigger the "trap" as soon as resuming the
program after a BP stop (or step). I.e. inhibit those part's when
debugging (which is probably most of the time). Remeber to re-enable for
"release" version of any application.

- Working on getting rid of all the compilation warnings.

- Detected a new serious bug. If an extra malloc is not executed
*before* the first thread is created that requires a steck  (i.e. the
idle tread sice root allready has a stack), that thread will fail with
an illegal OP-code trap. This has been traced to be due to a faulty
malloc and/or possibly a memory alignement problem. The first block
allocated, will be about 6 positions to high up in the memory map, which
means that sthe total block will not really fit. If that block is the
stack of a thread, those positions will be either the context or the
return adress of that thread (which is bad). I'm concerned about not
detecting this error before, which leads me to believe that this
actually is an alignement issue in malloc and it's anly pure chance
wheather the bug will manifest or not. This is a problem related
to the Keil_XC167 target only.