Build system and source modified to make better use of config.h
This file now contains information about how the kernel is configured
and can be used by both application and kernel build (old solution only
let kernel-buils know of these details).

This applies to both tk_tuning, component configuration among others.
Use './configure --help' to see a full list. Note that  if a certain
feature is not configured, the old tk_tuning will fill in the gaps.
This is especially usefull when not using GNU build- and configure-
tool-chain. Hopefully, we'll be able to get rid of tk_tuning.h in the
future.