Trying to allocate a better TOS
Filed but an embryo to solution might be there.

We might need linker support to get a root stack and/or use other means for
kalloc (which is mapping to malloc a.t.m.).

Notes:

HIXS_sbrk needs review. Is using the linkers _end symbol really correct for
_sbrk? This worked for PPC but we had much more memory and several sections
ther. For ARM linker script, this is on top of memory and the algorithm seems
to search upwards.