Build-system overhaul
64-bit build PC:s, whom are almost all multi architectural, don't place
glibc headers and libraries in a uniform way any longer to be able to
distinguish 32-bit from 64-bit versions. This creates a problem for us using
BUILDCHAIN wrapper method as this assumes there's only one path to the include and
library directories. When cross-compiling this is a non-issue as even if the
target is multi architectural, paths are handled differently.

Issue solved and mechanism, however: Names in a wrapped path that have
been defined will expand. For example:

/usr/include/x86_64-linux-gnu will expand to /usr/include/x86_64-1-gnu,
rendering include files not found. This is a tricky issue as the strings are
not really strings, which also explains why #include BUILDCHAIN(path) don't
take any surrounding " or <>. The preprocessor knows it's a string
implicitly due to it's # expansion and allows the construct, but we can't do
anything to manipulate or restrain it further.

Issue will be solved by identifying which macros are defined (easy to
determine) and undefine them (possibly temporarily if it turns out to
interfere with the tools or system).