Finally builds for 64bit hosted linux-target
First time in ~10 years we've been in any hosted mode. A shit-load has
happened on the CPU front. uC and CPU:s have diverged and it's probably not
a very effective principle to store each context on top of each stack, at
least not without some thought (consider preemption which will recurse
the stack).

However, TinKer is first an foremost a deeply embedded kernel. Almost equally
important as embeddedness is transparency and portability. It's still
possible to port TinKer to a new architecture in a day, even by this ol guy.
That is made possible only by not using specialized HW supported features.

Even without fancy MMU, SMP and specialized context-stacks it still beats
linux silly on pure performance any-day for the executables it's made for
though. That's what this port (_x86_64) aims at proving.

Do we dare to fire it up? Nah, lets wait and gather courage.

Remaining issues:
More registers need context saving.
* r8-r15
* xmm8-xmm15

Might not be needed since xmm0-xmm4 wasn't saved last time it run on a
ancestor to x86, I'm sure it must had been a PII at least :-). Our threads
are not going to do streaming but should the context be required to use
these, dynamic detection of usage is a must as they are riotously large
(16x16 bytes).