-DARCH=$(ARCH) for native builds changed [IMPORTANT]
This change is flagged [IMPORTANT] as it might have future impact which I
can't foresee right now.

Used to be -DARCH=_$(ARCH), not sure if this was a typo or an intentional
"reminder to self" for some reason. It was a very long time since hosted
tinker builds were executing.

Remark:
=======
It would be better if as many defines got passed via
./include/tinker/configure instead, but for legacy and backwards
comparability to non-GNU tool-chains it seems some -D are still passed this
way. But GNU has it's own set of make-files so it seems somewhat redundant.

The only relevant reason for having macros passed via -D is when to resolve
path-names to include-files with more macros. But configure.h is in a
well-known hard-coded place (relative) place so it should be possible to
remove at least some -Dxxx macros. BOARD and MCPU might be different, but
even here I'm uncertain of if the  build-system wont allow you to configure
with one BORD/MCPU and build with another, or if that would even be
desirable.

Remaining issues:
* As the configure system ban be 1) tricked by the build system, 2)
  vice versa and 3) since many macros have #ifndef macro fall-backs, we
  should have a runt-time query API with some of the more important
  build/configure settings returned in a static struct.
(cherry picked from commit 8550bd61069873bc2a8cbda5e42c8f3adbbff4e7)