Configure system: make.subdirs files removed
Dependency to "make.subdirs" removed as it's only confusing when one
build-system is used to configure another and what to configure is hidden in
yet another obscure file. When adding new platforms there will be a caos of
what is where and just because some parts in the build-system works doesn't
have to mean corresponding parts in the configure system works. *Too much*
flexibility can be counterproductive i.e.

The purpose of having those files in the first place were to help close down
parts of the tree where the configure-system was either broken or under
development. This feature still exists, but a better way to handle it would
be by using branches in the SCM and not merge on main until at both
build/configure works.

However to deploy the old feature: Just add a file "make.subdirs" in a
subdirectory and those directories will be pre-configured with "make
configure". This is really just a build-system boot-strap to create a
recursive set of configure scripts.  The real configuration is done with the
produced ./configure script, at any level even though it's recommended to
start from the root-level.
(cherry picked from commit 68c15a535d3faa0973158e69e28f1b33dbcd02a5)

Conflicts:
	bsp/gnu_arm/make.subdirs
	bsp/make.subdirs