Revert 1/3 "configure system: Many autohell errors fixed"
This part reverted:
    * Most changes are due to the following autoconf issue:
    http://www.gnu.org/software/autoconf/manual/autoconf.html#Expanded-Before-Required

      It doesn't say in a way a human would understand it, but apparently you
      can't have AC_PROG_CC in a macro. AC_REQUIRE it inside the macro doesn't
      help. The only choice I saw was to lift it out of the macros using it and
      expand it once before outside. Unfortunately this means many small changes
      in all configure.in files.

Reason is that it borks configure so that it fail fails several but not all (!)
AC_CONFIG, but not all, after a certain depth (bsp/gnu_arm/BOARD &
bsp/gnu_arm/MCPU).

Reason is still unknown. Autoconf however does not complain on 2.50 which was
the reason for this change in the first place.

Patching the following by hand in Makefiles-gnu would allow build to continue:

XCOMPILE       = 1
ARCH           = arm
ABI            = eabi

Note that the following are OK, especially board which the build later
complains about. That is because the define is never past men XCOMPILE is
false.:

BOARD          = stm32
MCPU           = cortex-m3