Corrected difftime and use there-of
This method is correct regardless signed integer overflow. That can only
have bearing if 1) compiler also can see the actual overflow 2) there would
be any practical reason for acting upon optimization. The only remaining
possible way a compiler can deem this to be needed, is doing so in run-time,
which would require a run-time executive below the kernel (language
dependent). If that ever happens, we will definitely have to know about it
as that integration will not come out of a box.

In case there is actually something beneath TinKer without mentioned
integration, it would have to be hosted-mode, in which case we let the
host's difftime worry about it instead