Race when creating new threads
If for example finalization of death is delayed due to pending join,
internal structures and state are vulnerable. There may be other use-cases.
This change aims to addresses all.

As tk_yield is called as part of thread creation everything after that point
that affects global consistency must be mistrusted and re-done.

An alternative solution could be to have a special yield that doesn't run
anything else until thread creation is final. But the same can happen for
example in true preemptive dispatch (i.e. context switch after IRQ) which we
have no control over. Enforcing no true dispatch would invalidate Tinker as
a true RT kernel that can react directly to external event as opposed to
indirectly by queuing in a run-queue. Direct reaction has faster response and
a simpler solution is better due to future changes (KISS). This solution is
therefore preferred.

Having both wouldn't hurt however as an unfinished thread isn't really ready
to run anyway and it would make the execution flow easier to follow and
debugging a somewhat less hard. Bug was hard to find and has been part
of the code for a while.

Bug found using test-case: test_ccSort.c (9e604fd).

Requirement for test: Enough threads to provoke the scenario.

Additional change in patch:
A simple detector for this kind of error.

Category: bug-fix