ARM: malloc requires different sbrk implementation.
This applies to non-hosted targets only (i.e. real physical targets) but
it's a conceptual change that's worth considering even for other
architectures.

The Newlib provided one depends on a linker-script which is incompatible as
the Newlib malloc assumes everything up until top-of memory is free.  The
top it does not know or care about only the bottom (i.e. from where to
start).

Tinker new memory layout  is actually traditional, but since we have a
defined position for the root-stack and since context or other important
things may lay at BOS we need to help avoiding that this doesn't get
overwritten.

Another possibility would be to put the root-stack below the heap in which
case heap growth couldn't touch it. However the reversed problem
(stack-smashing) would still be an issue. Possibly somewhat detectable
during context switch though.