License headers corrected
New published work and old under src/. I.e. not bug-fixes. Files in bsp/
remain "as-is" even though their headers wrt the year(s) are not correct
according to the rule of "first published" and will be changed back
(re-created from log).

Original changes are in the git-log and the rcs/cvs-log still kept in each
file. Copy remains in Git even if rcs-logs are removed.